+++ pjrt/devices_test.go
@@ -14,7 +14,7 @@ func TestClient_Devices(t *testing.T) {
 	require.NoErrorf(t, err, "Failed to create a client on %s", plugin)
 	fmt.Printf("\t%s\n", client)
 
-	devices, err := client.Devices()
+	devices, err := client.AllDevices()
 	require.NoError(t, err, "Failed to list devices for %s", client)
 
 	addressableDevices := client.AddressableDevices()