+++ docs/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 # Next
 
-- Added TPU installation.
+- `gopjrt_installer`: Added TPU installation.
+- Package `pjrt`:
+  - Renamed `Client.Devices` to `Client.AllDevices`.
+  - Added `Client.NumDevices` to get the number of addressable devices.
 
 # v0.8.5 2025/10/27: Fixed `xlabuilder` for newer C compilers.